% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{first_norm}
\alias{first_norm}
\title{First norm}
\usage{
first_norm(A)
}
\arguments{
\item{A}{numeric matrix}
}
\value{
Returns a number which is the first norm of the given matrix.
}
\description{
Function calculates the first norm of the given matrix.
}
\examples{
B <- matrix(rnorm(16))
first_norm(B)

}
