% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{frobenius_norm}
\alias{frobenius_norm}
\title{Frobenius norm}
\usage{
frobenius_norm(A)
}
\arguments{
\item{A}{numeric matrix}
}
\value{
Returns a number which is the frobenius norm of the given matrix.
}
\description{
Function calculates the frobenius norm of the given matrix.
}
