% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{inf_norm}
\alias{inf_norm}
\title{Infinity norm}
\usage{
inf_norm(A)
}
\arguments{
\item{A}{numeric matrix}
}
\value{
Returns a number which is the infinity norm of the given matrix.
}
\description{
Function calculates the infinity norm of the given matrix.
}
\examples{
B <- matrix(rnorm(16))
inf_norm(B)

}
