% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{lu_decomposition}
\alias{lu_decomposition}
\title{LU decomposition}
\usage{
lu_decomposition(A)
}
\arguments{
\item{A}{numeric matrix}
}
\value{
Function returns a list of L,U matrices, being an LU decomposition of matrix A.
}
\description{
Funcion implements the LU decomposition od a sqare matrix.
}
\examples{
B <- matrix(rnorm(16), nrow = 4, ncol = 4)
res <- lu_decomposition(B)
print(res)

}
