% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\docType{package}
\name{matricesPackage-package}
\alias{matricesPackage}
\alias{matricesPackage-package}
\title{matricesPackage: matricesPackage}
\description{
EN:This package implements a few basic matrix functions in R: LU and Cholesky decompositions, matrix determinant and a few matrix norms. Package was created for educational purposes. PL:Niniejszy pakiet implementuje kilka podstawowych funkcji związanych z macierzami, czyli Rozkłady LU, Choleskiego-Banachiewicza, kilka norm macierzowych oraz wyznacznika macierzy. Pakiet powstał w celach edukacyjnych.
}
\author{
\strong{Maintainer}: Filip Labus \email{wewarszawie@gmail.com}

Authors:
\itemize{
  \item Filip Labus \email{wewarszawie@gmail.com}
}

}
