% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{llt_decomposition}
\alias{llt_decomposition}
\title{Cholesky-Banachiewicz decopmposition}
\usage{
llt_decomposition(A)
}
\arguments{
\item{A}{numeric matrix}
}
\value{
Returns an lower-triangular matrix L after the LL^T decomposition.
}
\description{
Function implements Cholesky-Banachiewicz decopmposition on square positive defiife matrix.
}
\examples{
B <- matrix(rnorm(16), nrow = 4, ncol = 4)
C <- B \%*\% t(B)
L <- llt_decomposition(C)
print(L)

}
