% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\docType{package}
\name{matricesPackage-package}
\alias{matricesPackage}
\alias{matricesPackage-package}
\title{matricesPackage: Pakiet z macierzami}
\description{
Niniejszy pakiet implementuje kilka podstawowych funkcji związanych z macierzami, czyli Rozkłady LU, Choleskiego-Banachiewicza, mnożenie przez wektor, oblicznie normy wektora oraz wyznacznika macierzy. Pakiet powstał w celach edukacyjnych.
}
\author{
\strong{Maintainer}: Filip Labus \email{wewarszawie@gmail.com}

Authors:
\itemize{
  \item Filip Labus \email{wewarszawie@gmail.com}
}

}
