% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{matrix_det}
\alias{matrix_det}
\title{Matrix determinant}
\usage{
matrix_det(A)
}
\arguments{
\item{A}{numeric matrix}
}
\value{
Functio returns a number which is  the determinant of given matrix.
}
\description{
Function calculates the matrix determinant using Cholesky-Banachiewicz decopmposition.
}
